Updated mode screens with plots and parameters
I added all the parameter inputs to the different modes. I added the functionality of the "submit data" button on all modes, such that a meaningless (sin wave) plot will appear when pressed and the text inputs are not empty. I also introduced a new "parameters.txt" file to the Python file and added it to the folder. This file is currently not interacted with and no data is sent or retrieved from it, so that's a to-do. For assignment 1, the plots can be placeholders, but I'm not sure about saving parameter data. Maybe Matthew can help me out with that.